backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
psycopg2-binary>=2.9.9
redis>=5.0.0
python-dotenv>=1.0.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.5
Pillow>=10.2.0
